requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_jwt_extended
flask_migrate
validators
python-dotenv
